add: use NA item name in exchange tickets if possible
