+++ src/scripts/utils/data-assets/parseTicketFile.mjs
@@ -1,28 +1,6 @@
 import { readFileYaml } from "@foxkit/node-util/fs-yaml";
 
-import { readFromCache } from "../atlasacademy/cache.mjs";
 import { parseTicketMonth } from "./parseTicketMonth.mjs";
-
-let niceItem = null;
-let itemIdMap = null;
-
-async function prepareNiceItem() {
-  if (niceItem == null) {
-    niceItem = await readFromCache("JP", "nice_item_lang_en.json");
-    if (!niceItem) {
-      throw new Error("Could not read nice_item_lang_en from cache");
-    }
-  }
-}
-
-async function prepareIdMap() {
-  if (itemIdMap == null) {
-    itemIdMap = await readFileYaml("assets/data/itemIdMap.yml");
-    if (!itemIdMap) {
-      throw new Error("Could not read assets/data/itemIdMap.yml");
-    }
-  }
-}
 
 const months = new Set([
   "Jan",
@@ -40,11 +18,7 @@ const months = new Set([
 ]);
 
 export async function parseTicketFile(filePath) {
-  const [rawData] = await Promise.all([
-    readFileYaml(filePath),
-    prepareNiceItem(),
-    prepareIdMap()
-  ]);
+  const rawData = await readFileYaml(filePath);
   if (!rawData) throw new Error(`Couldn't read Prism Shop file`);
   const parsedData = new Object();
 
@@ -59,10 +33,7 @@ export async function parseTicketFile(filePath) {
     if (rawData[month].length !== 3) {
       throw new TypeError(`Expected property ${month} to contain 3 elements`);
     }
-    parsedData[month] = parseTicketMonth(rawData[month], {
-      niceItem,
-      itemIdMap
-    });
+    parsedData[month] = await parseTicketMonth(rawData[month]);
   }
 
   if (Object.keys(parsedData).length < 1) {
